+++ src/Widgets/Buttons/AdminButton.php
@@ -155,7 +155,12 @@ CSS;
 }
 
 .admin {
-    display: none
+    display: none;
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%
 }
 
 .is-admin .admin {