+++ tests/Search/SearchTest.php
@@ -44,6 +44,11 @@ class WeightedActiveRecord extends ActiveRecord implements SearchableInterface
         return false;
     }
 
+    public function getPermissionName(): string
+    {
+        return 'test';
+    }
+
     public function getSearchAttributes(): array
     {
         return ['name'];